chore: refactor types files
