+++ src/types/commits.ts
@@ -1,5 +1,3 @@
-import { SemverLevelNone } from './version';
-
 /**
  * A commit in repository
  */
@@ -29,3 +27,15 @@ export type CommitsSummary = {
   references: string[];
   nonConventional: string[];
 };
+
+export type TagNotes = {
+  tagName: string;
+  version: string;
+  releaseNotes: string;
+  changesDetected: Commit[];
+  existingTag: boolean;
+};
+
+export type SemverLevel = 'patch' | 'minor' | 'major';
+
+export type SemverLevelNone = SemverLevel | 'none';
