+++ src/types/index.ts
@@ -1,3 +1,2 @@
 export * from './commits';
-export * from './version';
 export * from './options';
